requirements.txt
aiofiles==0.7.0
anyio==3.3.1
asgiref==3.4.1
asyncio==3.4.3
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
ecdsa==0.17.0
fastapi==0.68.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==1.0.0b0
idna==3.2
itsdangerous==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
